Re-factor a bit of the header piece.
