+++ src/javascript/components/SitemapPanelHeader/SitemapPanelHeader.scss
@@ -1,0 +1,4 @@
+.header {
+    background: #fff;
+    overflow: overlay;
+}
